pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "jhonatas",email = "jhonatasassumpcao@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "joblib (>=1.5.1,<2.0.0)",
    "flask-pydantic-spec (>=0.8.5,<0.9.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "scikit-learn (==1.6.1)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
httpie = "^3.2.4"
isort = "^6.0.1"
pip-audit = "^2.9.0"
pydocstyle = "^6.3.0"
blue = "^0.9.1"
